Backend/requirements.txt
pika==1.3.0
pymongo==4.2.0
python-dotenv==0.20.0
